requirements.txt
torch>=1.9.0
numpy>=1.21.0
scipy>=1.7.0
matplotlib>=3.4.0
seaborn>=0.11.0
pyyaml>=5.4.0
gym>=0.21.0
stable-baselines3>=1.5.0
opencv-python>=4.5.0
tqdm>=4.62.0
pandas>=1.3.0
scikit-learn>=1.0.0
jupyter>=1.0.0
pytest>=6.2.0
black>=21.0.0
flake8>=3.9.0 